[C4-01] move DropdownWithLabel into dropdown component
