+++ src/containers/CreateGroupPage/CreateGroupPage.js
@@ -4,8 +4,7 @@ import { withRouter } from "react-router-dom";
 
 import { Header, Subheader } from "../../components/Typography";
 import { Button } from "../../components/Button";
-
-import { DropdownWithLabel } from "./components/DropdownWithLabel";
+import { DropdownWithLabel } from "../../components/Dropdown";
 
 import { LARGE_DEVICE_SIZE } from "../../global/const";
 import { useWindowDimensions } from "../../global/utils";
